requirements.txt
Flask[async]==2.0.1
python-dotenv==1.0.1
Werkzeug==2.0.1
gunicorn==20.1.0